+++ packages/sdk/changelog.md
@@ -4,6 +4,7 @@
 
 - refactor: simpler connection logic (deleting `reconnect`)
 - refactor: improved generics in types using contexts
+- bug: getReplica only accepts domain number
 
 ### 2.0.0-rc.4